app/requirements.txt
anyio==4.11.0
certifi==2025.10.5
dotenv==0.9.9
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mariadb==1.1.14
openpyxl==3.1.5
packaging==25.0
python-dotenv==1.1.1
python-telegram-bot==22.5
sniffio==1.3.1